Pinger

+++ solution.py
@@ -131,7 +131,7 @@ def ping(host, timeout=1):
 		time.sleep(1)  # one second
 		# array=[i]
 		# array[i]=delay #first loop at 1, then 2, then 3 ...
-		array.insert(i,delay*1000)
+		array.insert(i,int(delay*1000))
 		# array.append(delay)
 
 	# Calculate vars values and return them
